Blocking inheritance for gRPC: simplified tests
