--- extensions/grpc/deployment/src/test/java/io/quarkus/grpc/server/services/BlockingExtendingExtendingTestService.java
@@ -1,7 +1,0 @@
-package io.quarkus.grpc.server.services;
-
-import io.quarkus.grpc.GrpcService;
-
-@GrpcService
-public class BlockingExtendingExtendingTestService extends BlockingBaseExtendingTestService {
-}
